+++ Templates/SharePoint-ADFS/dsc/ConfigureSPSE.ps1
@@ -23,12 +23,12 @@ configuration ConfigureSPVM
         [Parameter(Mandatory)] [System.Management.Automation.PSCredential]$SPSuperReaderCreds
     )
 
-    Import-DscResource -ModuleName ComputerManagementDsc -ModuleVersion 9.0.0 # Custom
+    Import-DscResource -ModuleName ComputerManagementDsc -ModuleVersion 9.1.0 # Custom
     Import-DscResource -ModuleName NetworkingDsc -ModuleVersion 9.0.0
-    Import-DscResource -ModuleName ActiveDirectoryDsc -ModuleVersion 6.3.0
+    Import-DscResource -ModuleName ActiveDirectoryDsc -ModuleVersion 6.4.0
     Import-DscResource -ModuleName xCredSSP -ModuleVersion 1.4.0
     Import-DscResource -ModuleName WebAdministrationDsc -ModuleVersion 4.1.0
-    Import-DscResource -ModuleName SharePointDsc -ModuleVersion 5.4.0
+    Import-DscResource -ModuleName SharePointDsc -ModuleVersion 5.5.0
     Import-DscResource -ModuleName DnsServerDsc -ModuleVersion 3.0.0
     Import-DscResource -ModuleName CertificateDsc -ModuleVersion 5.1.0
     Import-DscResource -ModuleName SqlServerDsc -ModuleVersion 16.5.0
@@ -59,8 +59,11 @@ configuration ConfigureSPVM
     [String] $SharePointBitsPath = Join-Path -Path $SetupPath -ChildPath "Binaries" #[environment]::GetEnvironmentVariable("temp","machine")
     [String] $SharePointIsoFullPath = Join-Path -Path $SharePointBitsPath -ChildPath "OfficeServer.iso"
     [String] $SharePointIsoDriveLetter = "S"
-    [String] $LDAPCPFileFullPath = Join-Path -Path $SetupPath -ChildPath "Binaries\LDAPCP.wsp"
     [String] $AdfsDnsEntryName = "adfs"
+    [String] $LdapcpSolutionName = "LDAPCPSE"
+    [String] $LdapcpSolutionId = "ff36c8cf-e510-42fc-8ba3-18af3c316aec"
+    [String] $LdapcpReleaseId = "latest"
+    [String] $LDAPCPFileFullPath = Join-Path -Path $SetupPath -ChildPath "Binaries\$LdapcpSolutionName.wsp"
 
     # SharePoint settings
     [String] $SPDBPrefix = "SPDSC_"
@@ -132,7 +135,7 @@ configuration ConfigureSPVM
         }
 
         # Enable TLS 1.2 - https://learn.microsoft.com/en-us/azure/active-directory/app-proxy/application-proxy-add-on-premises-application#tls-requirements
-        # It's a best practice, and mandatory with Windows 2012 R2 (SharePoint 2013) to allow xRemoteFile to download releases from GitHub: https://github.com/PowerShell/xPSDesiredStateConfiguration/issues/405           
+        # This allows xRemoteFile to download releases from GitHub: https://github.com/PowerShell/xPSDesiredStateConfiguration/issues/405
         Registry EnableTLS12RegKey1 {
             Key = 'HKLM:\SYSTEM\CurrentControlSet\Control\SecurityProviders\SCHANNEL\Protocols\TLS 1.2\Client'; ValueName = 'DisabledByDefault'; ValueType = 'Dword'; ValueData = '0'; Ensure = 'Present' 
         }
@@ -181,17 +184,14 @@ configuration ConfigureSPVM
         }
 
         # Set registry keys to allow OneDrive NGSC to connect to SPS using OIDC - part 1 (machine-wide)
-        Registry PrioritizeOIDCOverLegacyAuthN {
+        Registry OneDriveOIDC_PrioritizeOIDCOverLegacyAuthN {
             Key = "HKLM:\SOFTWARE\Policies\Microsoft\OneDrive"; ValueName = "SharePointOnPremOIDC"; ValueType = "DWORD"; ValueData = "1"; Ensure = "Present" 
         }
-        Registry PrioritizeSPSOverSPO {
+        Registry OneDriveOIDC_PrioritizeSPSOverSPO {
             Key = "HKLM:\SOFTWARE\Policies\Microsoft\OneDrive"; ValueName = "SharePointOnPremPrioritization"; ValueType = "DWORD"; ValueData = "1"; Ensure = "Present" 
         }
-        Registry SetSPSUrl {
+        Registry OneDriveOIDC_SPSUrl {
             Key = "HKLM:\Software\Policies\Microsoft\OneDrive"; ValueName = "SharePointOnPremFrontDoorUrl"; ValueType = "String"; ValueData = "https://$SharePointSitesAuthority.$DomainFQDN"; Ensure = "Present" 
-        }
-        Registry SetOneDriveName {
-            Key = "HKLM:\Software\Policies\Microsoft\OneDrive"; ValueName = "SharePointOnPremTenantName"; ValueType = "String"; ValueData = "$SharePointSitesAuthority in $DomainFQDN"; Ensure = "Present" 
         }
         
         SqlAlias AddSqlAlias {
@@ -201,7 +201,13 @@ configuration ConfigureSPVM
         Script EnableFileSharing {
             GetScript  = { }
             TestScript = { return $null -ne (Get-NetFirewallRule -DisplayGroup "File And Printer Sharing" -Enabled True -ErrorAction SilentlyContinue | Where-Object { $_.Profile -eq "Domain" }) }
-            SetScript  = { Set-NetFirewallRule -DisplayGroup "File And Printer Sharing" -Enabled True -Profile Domain -Confirm:$false }
+            SetScript  = { Set-NetFirewallRule -DisplayGroup "File And Printer Sharing" -Enabled True -Profile Domain }
+        }
+
+        Script EnableRemoteEventViewerConnection {
+            GetScript  = { }
+            TestScript = { return $null -ne (Get-NetFirewallRule -DisplayGroup "Remote Event Log Management" -Enabled True -ErrorAction SilentlyContinue | Where-Object { $_.Profile -eq "Domain" }) }
+            SetScript  = { Set-NetFirewallRule -DisplayGroup "Remote Event Log Management" -Enabled True -Profile Domain }
         }
 
         # Create the rules in the firewall required for the distributed cache - https://learn.microsoft.com/en-us/sharepoint/administration/plan-for-feeds-and-the-distributed-cache-service#firewall
@@ -220,18 +226,19 @@ configuration ConfigureSPVM
                 }
             }
             SetScript  = {
-                $icmpRuleName = "File and Printer Sharing (Echo Request - ICMPv4-In)"
-                $icmpFirewallRule = Get-NetFirewallRule -DisplayName $icmpRuleName -ErrorAction SilentlyContinue
-                if ($null -eq $icmpFirewallRule) {
-                    New-NetFirewallRule -Name Allow_Ping -DisplayName $icmpRuleName `
-                        -Description "Allow ICMPv4 ping" `
-                        -Protocol ICMPv4 `
-                        -IcmpType 8 `
-                        -Enabled True `
-                        -Profile Any `
-                        -Action Allow
-                }
-                Enable-NetFirewallRule -DisplayName $icmpRuleName
+                # $icmpRuleName = "File and Printer Sharing (Echo Request - ICMPv4-In)"
+                # $icmpFirewallRule = Get-NetFirewallRule -DisplayName $icmpRuleName -ErrorAction SilentlyContinue
+                # if ($null -eq $icmpFirewallRule) {
+                #     New-NetFirewallRule -Name Allow_Ping -DisplayName $icmpRuleName `
+                #         -Description "Allow ICMPv4 ping" `
+                #         -Protocol ICMPv4 `
+                #         -IcmpType 8 `
+                #         -Enabled True `
+                #         -Profile Any `
+                #         -Action Allow
+                # }
+                # Enable-NetFirewallRule -DisplayName $icmpRuleName
+                Enable-NetFirewallRule -displayName "File and Printer Sharing (Echo Request - ICMPv4-In)"
 
                 $spRuleName = "SharePoint Distributed Cache"
                 $firewallRule = Get-NetFirewallRule -DisplayName $spRuleName -ErrorAction SilentlyContinue
@@ -249,7 +256,7 @@ configuration ConfigureSPVM
 
         xRemoteFile DownloadLDAPCP {
             DestinationPath = $LDAPCPFileFullPath
-            Uri             = Get-LatestGitHubRelease -Repo "Yvand/LDAPCP" -Artifact "*.wsp" -ReleaseId "latest"
+            Uri             = Get-LatestGitHubRelease -Repo "Yvand/LDAPCP" -Artifact "*.wsp" -ReleaseId $LdapcpReleaseId
             MatchSource     = $false
         }
 
@@ -437,6 +444,32 @@ configuration ConfigureSPVM
             Name = "Default Web Site"; Ensure = "Absent"; PhysicalPath = "C:\inetpub\wwwroot"; 
         }
 
+        Script CreateSPLOGSFileShare {
+            SetScript  = 
+            { 
+                $foldername = "C:\Program Files\Common Files\Microsoft Shared\Web Server Extensions\16\LOGS"
+                $shareName = "SPLOGS"
+                # if (!(Get-WmiObject Win32_Share -Filter "name='$sharename'")) {
+                $shares = [WMICLASS]"WIN32_Share"
+                if ($shares.Create($foldername, $sharename, 0).ReturnValue -ne 0) {
+                    Write-Host "Failed to create file share '$sharename' for folder '$foldername'"
+                } else {
+                    Write-Host "Created file share '$sharename' for folder '$foldername'"
+                }
+                # }
+            }
+            GetScript  = { }
+            TestScript = 
+            {
+                $shareName = "SPLOGS"
+                if (!(Get-WmiObject Win32_Share -Filter "name='$sharename'")) {
+                    return $false
+                } else {
+                    return $true
+                }
+            }
+        }
+
         #**********************************************************
         # Join AD forest
         #**********************************************************
@@ -514,12 +547,16 @@ configuration ConfigureSPVM
         }
 
         # Set registry keys to allow OneDrive NGSC to connect to SPS using OIDC - part 2 (user-specific settings)
-        Registry EnableOIDCForSPS {
+        Registry OneDriveOIDC_EnableOIDCForSPS {
             Key = "HKCU:\Software\Microsoft\OneDrive\PreSignInRampOverrides"; ValueName = "2086"; ValueType = "DWORD"; ValueData = "1"
             PsDscRunAsCredential = $DomainAdminCredsQualified; Ensure = "Present" 
         }
-        Registry EnableOneAuthorSPS {
+        Registry OneDriveOIDC_EnableOneAuthorSPS {
             Key = "HKCU:\Software\Microsoft\OneDrive\PreSignInRampOverrides"; ValueName = "2042"; ValueType = "DWORD"; ValueData = "1"
+            PsDscRunAsCredential = $DomainAdminCredsQualified; Ensure = "Present" 
+        }
+        Registry OneDriveOIDC_IniFileNamingForEmailAndUpn {
+            Key = "HKCU:\Software\Microsoft\OneDrive\PreSignInSettingsOverrides"; ValueName = "204"; ValueType = "String"; ValueData = "-1835"
             PsDscRunAsCredential = $DomainAdminCredsQualified; Ensure = "Present" 
         }
 
@@ -810,11 +847,8 @@ configuration ConfigureSPVM
         Script RestartSPTimerAfterCreateSPFarm {
             SetScript            =
             {
-                # Restarting SPTimerV4 service before deploying solution makes deployment a lot more reliable
-                Restart-Service SPTimerV4
-                # 2021-09: In SharePoint 2013, solution deployment failed multiple times with error "Admin SVC must be running in order to create deployment timer job."
-                # So ensure that SPAdminV4 is started
-                Restart-Service SPAdminV4
+                # Restarting both SPAdminV4 and SPTimerV4 services before deploying solution makes deployment a lot more reliable
+                Restart-Service SPTimerV4, SPAdminV4
             }
             GetScript            = { }
             TestScript           = { return $false } # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
@@ -841,13 +875,38 @@ configuration ConfigureSPVM
             DependsOn            = "[File]CopyCertificatesFromDC"
         }
 
-        SPFarmSolution InstallLdapcp {
+        SPFarmSolution InstallLdapcpSolution {
             LiteralPath          = $LDAPCPFileFullPath
-            Name                 = "LDAPCP.wsp"
+            Name                 = "$LdapcpSolutionName.wsp"
             Deployed             = $true
             Ensure               = "Present"
             PsDscRunAsCredential = $SPSetupCredsQualified
             DependsOn            = "[Script]RestartSPTimerAfterCreateSPFarm"
+        }
+
+        Script InstallLdapcpFeatures {
+            SetScript            = 
+            {
+                $solutionId = $using:LdapcpSolutionId
+                Install-SPFeature -SolutionId $solutionId -AllExistingFeatures
+            }
+            GetScript            =  
+            {
+                # This block must return a hashtable. The hashtable must only contain one key Result and the value must be of type String.
+                return @{ "Result" = "false" }
+            }
+            TestScript           = 
+            {
+                # If it returns $false, the SetScript block will run. If it returns $true, the SetScript block will not run.
+                $claimsProviderName = $using:LdapcpSolutionName
+                if ($null -eq (Get-SPClaimProvider -Identity $claimsProviderName -ErrorAction SilentlyContinue)) {
+                    return $false
+                } else {
+                    return $true
+                }
+            }
+            DependsOn            = "[SPFarmSolution]InstallLdapcpSolution"
+            PsDscRunAsCredential = $DomainAdminCredsQualified
         }
 
         SPManagedAccount CreateSPSvcManagedAccount {
@@ -966,28 +1025,28 @@ configuration ConfigureSPVM
             PsDscRunAsCredential = $DomainAdminCredsQualified
         }
 
-        # Installing LDAPCP somehow updates SPClaimEncodingManager 
-        # But in SharePoint 2019 and Subscription, it causes an UpdatedConcurrencyException on SPClaimEncodingManager in SPTrustedIdentityTokenIssuer resource
-        # The only solution I've found is to force a reboot
-        Script ForceRebootBeforeCreatingSPTrust {
-            # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
-            TestScript           = {
-                return (Test-Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust)
-            }
-            SetScript            = {
-                New-Item -Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust -Force
-                $global:DSCMachineStatus = 1
-            }
-            GetScript            = { }
-            PsDscRunAsCredential = $DomainAdminCredsQualified
-            DependsOn            = "[SPFarmSolution]InstallLdapcp"
-        }
+        # # Installing LDAPCP somehow updates SPClaimEncodingManager 
+        # # But in SharePoint 2019 and Subscription, it causes an UpdatedConcurrencyException on SPClaimEncodingManager in SPTrustedIdentityTokenIssuer resource
+        # # The only solution I've found is to force a reboot
+        # Script ForceRebootBeforeCreatingSPTrust {
+        #     # If the TestScript returns $false, DSC executes the SetScript to bring the node back to the desired state
+        #     TestScript           = {
+        #         return (Test-Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust)
+        #     }
+        #     SetScript            = {
+        #         New-Item -Path HKLM:\SOFTWARE\DscScriptExecution\flag_ForceRebootBeforeCreatingSPTrust -Force
+        #         $global:DSCMachineStatus = 1
+        #     }
+        #     GetScript            = { }
+        #     PsDscRunAsCredential = $DomainAdminCredsQualified
+        #     DependsOn            = "[SPFarmSolution]InstallLdapcpSolution"
+        # }
 
-        PendingReboot RebootOnSignalFromForceRebootBeforeCreatingSPTrust {
-            Name             = "RebootOnSignalFromForceRebootBeforeCreatingSPTrust"
-            SkipCcmClientSDK = $true
-            DependsOn        = "[Script]ForceRebootBeforeCreatingSPTrust"
-        }
+        # PendingReboot RebootOnSignalFromForceRebootBeforeCreatingSPTrust {
+        #     Name             = "RebootOnSignalFromForceRebootBeforeCreatingSPTrust"
+        #     SkipCcmClientSDK = $true
+        #     DependsOn        = "[Script]ForceRebootBeforeCreatingSPTrust"
+        # }
 
         $apppoolUserName = $SPAppPoolCredsQualified.UserName
         $domainAdminUserName = $DomainAdminCredsQualified.UserName
@@ -1047,7 +1106,7 @@ configuration ConfigureSPVM
                     return $true
                 }
             }
-            DependsOn            = "[SPFarmSolution]InstallLdapcp"
+            DependsOn            = "[Script]RestartSPTimerAfterCreateSPFarm"
             PsDscRunAsCredential = $DomainAdminCredsQualified
         }
 
@@ -1074,11 +1133,11 @@ configuration ConfigureSPVM
                 }
             )
             SigningCertificateFilePath = "$SetupPath\Certificates\ADFS Signing.cer"
-            ClaimProviderName          = "LDAPCP"
+            ClaimProviderName          = $LdapcpSolutionName
             UseWReplyParameter         = $true
             Ensure                     = "Present" 
-            DependsOn                  = "[Script]SetFarmPropertiesForOIDC"
-            PsDscRunAsCredential       = $SPSetupCredsQualified
+            DependsOn                  = "[Script]SetFarmPropertiesForOIDC", "[Script]InstallLdapcpFeatures"
+            PsDscRunAsCredential       = $DomainAdminCredsQualified
         }
 
 
@@ -1099,28 +1158,8 @@ configuration ConfigureSPVM
         Script ConfigureLDAPCP {
             SetScript            = 
             {
-                Add-Type -AssemblyName "ldapcp, Version=1.0.0.0, Culture=neutral, PublicKeyToken=80be731bc1a1a740"
-
-                # Create LDAPCP configuration
-                $config = [ldapcp.LDAPCPConfig]::CreateConfiguration([ldapcp.ClaimsProviderConstants]::CONFIG_ID, [ldapcp.ClaimsProviderConstants]::CONFIG_NAME, $using:DomainFQDN);
-
-                # LDAP://contoso.local:636/CN=Users,DC=contoso,DC=local
-
-                # Remove unused claim types
-                $config.ClaimTypes.Remove("http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress")
-                $config.ClaimTypes.Remove("http://schemas.microsoft.com/ws/2008/06/identity/claims/windowsaccountname")
-                $config.ClaimTypes.Remove("http://schemas.microsoft.com/ws/2008/06/identity/claims/primarygroupsid")
-                $config.ClaimTypes.Remove("http://schemas.microsoft.com/ws/2008/06/identity/claims/role")
-                
-                # Configure augmentation
-                $config.EnableAugmentation = $true
-                $config.MainGroupClaimType = "http://schemas.microsoft.com/ws/2008/06/identity/claims/role"
-                foreach ($connection in $config.LDAPConnectionsProp) {
-                    $connection.EnableAugmentation = $true
-                }
-
-                # Save changes
-                $config.Update()
+                Add-Type -AssemblyName "Yvand.LDAPCPSE, Version=1.0.0.0, Culture=neutral, PublicKeyToken=80be731bc1a1a740"
+                [Yvand.LdapClaimsProvider.LDAPCPSE]::CreateConfiguration()
             }
             GetScript            =  
             {
@@ -1130,8 +1169,8 @@ configuration ConfigureSPVM
             TestScript           = 
             {
                 # If it returns $false, the SetScript block will run. If it returns $true, the SetScript block will not run.
-                Add-Type -AssemblyName "ldapcp, Version=1.0.0.0, Culture=neutral, PublicKeyToken=80be731bc1a1a740"
-                $config = [ldapcp.LDAPCPConfig]::GetConfiguration("LDAPCPConfig")
+                Add-Type -AssemblyName "Yvand.LDAPCPSE, Version=1.0.0.0, Culture=neutral, PublicKeyToken=80be731bc1a1a740"
+                $config = [Yvand.LdapClaimsProvider.LDAPCPSE]::GetConfiguration()
                 if ($config -eq $null) {
                     return $false
                 }
@@ -1331,7 +1370,7 @@ configuration ConfigureSPVM
             ApplicationPool      = $ServiceAppPoolName
             DatabaseName         = "$($SPDBPrefix)SubscriptionSettings"
             PsDscRunAsCredential = $SPSetupCredsQualified
-            DependsOn            = "[SPServiceAppPool]MainServiceAppPool", "[SPServiceInstance]StartSubscriptionSettingsServiceInstance", "[Script]ConfigureLDAPCP"
+            DependsOn            = "[SPServiceAppPool]MainServiceAppPool", "[SPServiceInstance]StartSubscriptionSettingsServiceInstance"
         }
 
         SPAppManagementServiceApp CreateAppManagementServiceApp {
@@ -1541,7 +1580,7 @@ configuration ConfigureSPVM
             DestinationPath = "C:\inetpub\wwwroot\addins"
             Type            = "Directory"
             Ensure          = "Present"
-            DependsOn       = "[SPFarm]CreateSPFarm", "[Script]ConfigureLDAPCP"
+            DependsOn       = "[SPFarm]CreateSPFarm"
         }
 
         WebAppPool CreateAddinsSiteApplicationPool {